requirements.txt
-e .
attrs==19.3.0
aws-cdk.assets==1.22.0
aws-cdk.aws-apigateway==1.22.0
aws-cdk.aws-certificatemanager==1.22.0
aws-cdk.aws-cloudformation==1.22.0
aws-cdk.aws-cloudwatch==1.22.0
aws-cdk.aws-ec2==1.22.0
aws-cdk.aws-elasticloadbalancingv2==1.22.0
aws-cdk.aws-events==1.22.0
aws-cdk.aws-iam==1.22.0
aws-cdk.aws-kms==1.22.0
aws-cdk.aws-lambda==1.22.0
aws-cdk.aws-logs==1.22.0
aws-cdk.aws-route53==1.22.0
aws-cdk.aws-s3==1.22.0
aws-cdk.aws-s3-assets==1.22.0
aws-cdk.aws-sns==1.22.0
aws-cdk.aws-sqs==1.22.0
aws-cdk.aws-ssm==1.22.0
aws-cdk.core==1.22.0
aws-cdk.cx-api==1.22.0
aws-cdk.region-info==1.22.0
cattrs==1.0.0
jsii==0.21.2
publication==0.0.3
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.1
